feat: Complete Gemma 3 4B multimodal fine-tuning on RTX 3070 8GB
- Add main training script: gemma_multimodal_qlora_finetuning.py
- Add continue training script: continue_training.py
- Add quick test script: quick_test.py
- Update comprehensive README.md with full documentation
- Update .gitignore for model files and cache
- Clean up project structure
- Achieve 89% loss reduction in 8 minutes training
- Successfully fine-tune 4B model on consumer 8GB GPU

+++ unsloth/unsloth_compiled_cache/Conv3d.py
@@ -56,4 +56,4 @@ from typing import Any, List, Optional, Tuple, Union, Dict, Set, Callable
 
 
 def forward(self, input: Tensor) -> Tensor:
-    return self._conv_forward(input, self.weight, self.bias).to(input.dtype).to(input.dtype).to(input.dtype)
+    return self._conv_forward(input, self.weight, self.bias).to(input.dtype).to(input.dtype)
